guest-voice-book/requirements.txt
python-statemachine
gpiozero
RPi.GPIO
pygame
pydub
